added maxModulus() for ease

+++ linbox/linbox/algorithms/signature.h
@@ -299,8 +299,8 @@ namespace LinBox
 		static Vector& symmetricLU (Vector& v, const Matrix& IM)
 		{
 
-			// typedef Modular<int32_t> Field;
-			typedef Modular<double> Field;
+			typedef Modular<int32_t> Field;
+			// typedef Modular<double> Field;
 			typedef Field::Element Element;
 			typedef DenseMatrix<Field> FMatrix;
 			RandomPrimeIterator primeg(20);
